% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiviste.R
\name{archiviste}
\alias{archiviste}
\title{L'Archiviste : A user-friendly database management}
\usage{
archiviste(
  database,
  additions,
  sample.ids,
  method = "jw",
  save.updates = TRUE,
  save.step = FALSE,
  save.report = TRUE,
  return.report = FALSE,
  save.dir,
  database.label,
  note,
  sep = ",",
  dec = ".",
  na.strings = "",
  na = "",
  fileEncoding = "latin1"
)
}
\arguments{
\item{database}{A data frame or a file path (character) to the existing database.}

\item{additions}{A data frame or a file path (character) to the additions.}

\item{sample.ids}{A character vector specifying the column names that uniquely identify samples.}

\item{method}{A character string specifying the matching method (default: `"jw"` for Jaro-Winkler).}

\item{save.updates}{Logical; if `TRUE`, saves the updated database (default: `TRUE`).}

\item{save.step}{Logical; if `TRUE`, saves intermediate update steps (default: `FALSE`).}

\item{save.report}{Logical; if `TRUE`, saves a report of updates made (default: `TRUE`).}

\item{return.report}{Logical; if `TRUE`, returns both the updated database and the report (default: `FALSE`).}

\item{save.dir}{Character. Optional directory path for saving the report. Required if `save.report = TRUE`.}

\item{database.label}{Character. Label for the database, used for naming output files.}

\item{note}{Character. Optional additional note to append to the file and report file name.}

\item{sep}{A character specifying the field separator for CSV files (default: `","`).}

\item{dec}{A character specifying the decimal separator for CSV files (default: `"."`).}

\item{na.strings}{A character vector specifying strings to be treated as `NA` (default: `""`).}

\item{na}{Character string, used for missing values in the data.= (default - "").}

\item{fileEncoding}{Character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded as they are written, "latin1" (default).}
}
\value{
If `return.report = TRUE`, returns a list with the updated database and a report. Otherwise, returns only the updated database.
}
\description{
This function updates an existing database with new additions, ensuring consistency in sample identifiers and columns.
It also generates a report detailing the updates made.
}
\examples{
\dontrun{
updated.db <- archiviste(
  database = "path.to/database.csv",
  additions = "path.to/additions.csv",
  sample.ids = c("Sample_name", "IGSN"),
  save.updates = TRUE,
  save.report = TRUE,
  save.dir = "output/directory"
)
}

}
\author{
Thomas Chalaux-Clergue
}
