% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservateur.R
\name{conservateur}
\alias{conservateur}
\title{Le Conversateur: Ensures that sample information is revised when necessary}
\usage{
conservateur(
  database,
  additions,
  sample.ids,
  save.updates = TRUE,
  save.report = TRUE,
  return.report = FALSE,
  save.dir,
  database.label,
  note,
  na = "",
  fileEncoding = "latin1"
)
}
\arguments{
\item{database}{Data frame. The original database containing sample information.}

\item{additions}{Data frame. New samples or updates of the existing database}

\item{sample.ids}{List or Character. The name(s) of the column(s) containing sample identifiers.}

\item{save.report}{Logical. If `TRUE`, a report summarizing the updates will be saved. Default is `TRUE`.}

\item{return.report}{Logical. If `TRUE`, the report is included in the function output. Default is `TRUE`.}

\item{save.dir}{Character. Optional directory path for saving the report. Required if `save.report = TRUE`.}

\item{database.label}{Character. Label for the database, used for naming output files.}

\item{note}{Character. Optional additional note to append to the file and report file name.}

\item{na}{Character string, used for missing values in the data.= (default - "").}

\item{fileEncoding}{Character string, if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded as they are written, "latin1" (default).}
}
\value{
A dataframe containing the updated database, and if the report is not save - a character vector of the report contents.
}
\description{
This function updates a database by resolving conflicts between existing samples and amended samples.
It allows users to decide whether to keep old values, replace them with new values, or apply decisions globally.
}
\details{
The function iterates through samples that exist in both `database` and `amended.samples` and compares column values.
If discrepancies are found, the user is prompted to choose whether to keep the old value, replace it with the new value,
or apply the decision to all similar cases.

A detailed report is generated, documenting the changes made.
}
\examples{
\dontrun{
updated.database <- conservateur(
  database = original.db,
  amended.samples = updated.samples,
  common.samples = intersect(original.db$Sample_name, updated.samples$Sample_name),
  sample.id = "Sample_name",
  save.report = TRUE,
  return.report = TRUE,
  save.dir = "./reports",
  database.label = "study_database",
  note = "revision1")
}

}
\author{
Thomas Chalaux-Clergue
}
