% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptioniste.R
\name{receptioniste}
\alias{receptioniste}
\title{Le receptioniste: Interactive CSV Loader}
\usage{
receptioniste(file, ..., db.type)
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{...}{Additional arguments to pass to `read.csv()`, such as `header`, `sep`, `dec`, `na.strings`, `fileEncoding`,}
}
\value{
A data frame if the file loads correctly; otherwise, the function continues prompting the user for corrections.
}
\description{
The `receptioniste` function is an interactive wrapper around `read.csv()` that helps users load a CSV file correctly.
If the file does not load properly (e.g., only one column is detected due to an incorrect separator),
the function prompts the user to modify parameters or add new arguments dynamically (see 'read.table \{utils\}' for more details).
}
\details{
This function helps users debug and adjust CSV import parameters interactively.
If the file loads incorrectly (only one column detected), users are prompted to:
- Modify existing `utils::read.csv()` arguments.
- Add new arguments dynamically.
- Change the file path if needed.
}
\examples{
\dontrun{
# Load a CSV file with default settings
database <- receptioniste("database.csv", sep = ";", fileEncoding = "latin1", na.strings = "NA")

# If the file doesn't load correctly, the user is prompted to modify parameters
}

}
\author{
Thomas Chalaux-Clergue
}
